mixin 적용
