+++ src/app/page.tsx
@@ -1,4 +1,5 @@
 //              src/app/page.tsx
+//              메인 교육
 export default function Home() {
   return (
    <main className="mainContainer">
