+++ src/components/footer/footer.scss
@@ -1,0 +1,4 @@
+
+footer{
+    background-color: $gray;
+}
