+++ src/components/header/header.scss
@@ -1,0 +1,33 @@
+@use '../../styles/main' as *;
+
+#mainHeader {
+    padding: 0 2rem;
+
+    .headerTOP {
+        @include flexbox();
+    }
+
+    .titleLogo {
+        margin: 2rem 0;
+    }
+
+    .logo {
+        display: flex;
+        justify-content: center;
+        font-size: 2rem;
+        text-align: center;
+        color: rgb(0, 202, 169);
+        font-family: 'TTHakgyoansimSamulhamR';
+    }
+
+    .subMenuBar {
+        display: flex;
+        justify-content: center;
+        align-items: center;
+        gap: 2rem;
+
+        .link {
+            text-decoration: none;
+        }
+    }
+}
