+++ src/essets/base/reset.scss
@@ -7,25 +7,47 @@
     padding: 0;
 }
 
+h1, h2, h3, h4, h5, h6 {
+  margin: 0;
+}
+
+a {
+  color: inherit;
+  text-decoration: none;
+}
+
+button,
+input,
+select,
+textarea {
+  background-color: transparent;
+  border: 0;
+
+  &:focus {
+    outline: none;
+    box-shadow: none;
+  }
+}
+
+a,
+button {
+  cursor: pointer;
+}
+
+ul,
+ol {
+  padding-left: 0;
+  list-style: none;
+}
+
 html {
     background-color: cadetblue;
 }
 
 body {
     margin: 0 auto;
-    padding: 0 2rem;
     background-color: #fff;
     width: 100%;
     max-width: 768px;
     height: 100svh;
-}
-
-.bodyContainer {
-    margin: 0 2rem;
-}
-
-ul,
-li {
-    outline: 0;
-    list-style: none;
 }
