+++ src/styles/constants/breakpoint.scss
@@ -1,0 +1,2 @@
+$md-breakpoint: 768px;
+$sm-breakpoint: 430px;
