+++ src/styles/main.scss
@@ -1,1 +1,14 @@
 @import './common.scss';
+
+@import './constants/breakpoint.scss';
+@import './constants/typography.scss';
+
+@import './mixins/flexbox.scss';
+@import './mixins/positions.scss';
+@import './mixins/responsive.scss';
+@import './mixins/textstyle.scss';
+
+@import './module/buttons.scss';
+
+@import './constants/breakpoint.scss';
+@import './constants/colors.scss';
