--- src/styles/module/Footer.module.css
@@ -1,3 +1,0 @@
-footer{
-    background-color: gray;
-}
