--- src/styles/module/Header.module.css
@@ -1,36 +1,0 @@
-   @font-face {
-       font-family: 'TTHakgyoansimSamulhamR';
-       src: url('https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2402_keris@1.0/TTHakgyoansimSamulhamR.woff2') format('woff2');
-       font-weight: normal;
-       font-style: normal;
-   }
-
-   #header {
-       width: 768px;
-       margin: 2rem;
-       height: 100px;
-
-       .titleLogo {
-           margin: 2rem 0;
-
-           .logo {
-               display: flex;
-               justify-content: center;
-               font-size: 2rem;
-               text-align: center;
-               color: rgb(0, 202, 169);
-               font-family: 'TTHakgyoansimSamulhamR';
-           }
-       }
-
-       .subMenuBar {
-           display: flex;
-           justify-content: center;
-           align-items: center;
-           gap: 2rem;
-
-           .link {
-               text-decoration: none;
-           }
-       }
-   }
