+++ src/styles/module/buttons.scss
@@ -1,0 +1,78 @@
+// %btn-base {
+//   @include inline-flexbox;
+//   padding: 0 8px;
+//   font-weight: 700;
+//   border-radius: 4px;
+
+//   &:disabled {
+//     cursor: not-allowed;
+//     opacity: 0.4;
+//   }
+// }
+
+// // Note: Button Styles
+// .btn-primary {
+//   @extend %btn-base;
+//   color: $white;
+//   background-color: $blue;
+//   transition: background-color 200ms ease-in-out;
+
+//   &:not(:disabled):hover {
+//     background-color: $blue-dark;
+//   }
+// }
+
+// .btn-secondary {
+//   @extend %btn-base;
+//   color: $primary;
+//   background-color: $border;
+//   transition: background-color 200ms ease-in-out;
+
+//   &:not(:disabled):hover {
+//     background-color: $tertiary;
+//   }
+// }
+
+// .btn-outlined {
+//   @extend %btn-base;
+//   color: $blue;
+//   background-color: $white;
+//   border: 1px solid blue;
+//   transition: background-color 200ms ease-in-out;
+
+//   &:not(:disabled):hover {
+//     background-color: $blue-light;
+//   }
+// }
+
+// .btn-ghost {
+//   @extend %btn-base;
+//   color: $primary;
+//   background-color: transparent;
+//   transition: color 200ms ease-in-out;
+
+//   &:not(:disabled):hover {
+//     color: $secondary;
+//   }
+// }
+
+// // Note: Button Sizes
+// .btn-32 {
+//   @include text-style(14);
+//   height: 32px;
+// }
+
+// .btn-40 {
+//   @include text-style(16);
+//   height: 40px;
+// }
+
+// .btn-48 {
+//   @include text-style(16);
+//   height: 48px;
+// }
+
+// .btn-55 {
+//   @include text-style(18);
+//   height: 55px;
+// }
